+++ package.json
@@ -53,6 +53,7 @@
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.5",
+    "husky": "^9.0.6",
     "lint-staged": "^15.2.0",
     "openapi-typescript": "^6.7.4",
     "openapi-typescript-codegen": "^0.27.0",
