+++ src/components/feedbacks/Feedbacks.module.scss
@@ -101,12 +101,19 @@ color: #567343;
         font-weight: 700;
         line-height: 1.3;
         text-transform: uppercase;
+        border: none;
         border-radius: 10px;
-        background: var(--Dark-Green, #567343)
+        background: #567343;
+
 }
 .submitWrapper {
     width: 100%;
     display: flex;
-    justify-content: space-between;
+    justify-content:space-evenly ;
     align-items: center;
+}
+.formContainer {
+   border-radius: 10px;
+   border: 1px solid #FEEEE1;
+   padding: 30px;
 }
