+++ src/pages/feedback-page/FeedbackPage.module.scss
@@ -5,3 +5,8 @@
    margin-bottom: 197px;
    padding: 0 40px;
 }
+.feedbacksWrapper {
+   display: flex;
+   justify-content: space-between;
+   
+}
