+++ src/pages/feedback-page/FeedbackPage.tsx
@@ -1,0 +1,19 @@
+
+import Title from '../../components/feedbacks/Title';
+import ListOfFeedbacks from '../../components/feedbacks/list/ListOfFeedbacks';
+import FeedbackForm from '../../components/feedbacks/form/FeedbackForm';
+import styles from './FeedbackPage.module.scss';
+
+const FeedbackPage = () => {
+  return (
+      <div className={styles.container}>
+     <Title name='Відгуки' />
+          <div className={styles.feedbacksWrapper}>
+              <ListOfFeedbacks />
+              <FeedbackForm/>
+          </div>    
+    </div>
+  )
+}
+
+export default FeedbackPage
